refactor: rename `user` module to convention's `users`
