+++ src/modules/auth/interfaces/auth.interface.ts
@@ -1,4 +1,4 @@
-import { IUser } from '../../user/interfaces/user.interface';
+import { IUser } from '../../users/interfaces/user.interface';
 
 export interface IAuthPayload {
   sub: number;
