+++ src/modules/users/user.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
-import { UsersService } from './services/user.service';
+import { UserService } from './services/user.service';
 import { UserRepository } from './repositories/user.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
-  providers: [UsersService, UserRepository],
-  exports: [UsersService],
+  providers: [UserService, UserRepository],
+  exports: [UserService],
 })
 export class UserModule {}
